Moving all classes into correct package based on Java conventions
