+++ java/src/main/java/com/bacheconsulting/kata/encode/SessionModificationCmd.java
@@ -1,4 +1,4 @@
-package encodeKata;
+package com.bacheconsulting.kata.encode;
 
 public class SessionModificationCmd {
     private final MessageType messageType;
