+++ java/src/main/java/com/bacheconsulting/kata/encode/YEW.java
@@ -1,4 +1,4 @@
-package encodeKata;
+package com.bacheconsulting.kata.encode;
 
 public enum YEW {
     YewXyzTimer(8),
